cyber-agent/requirement.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo_search
fastapi
uvicorn
groq 
openai